feat(Root): Move channel item in new file `Channel.jsx`
